+++ packages/fx-core/src/plugins/resource/simpleauth/index.ts
@@ -18,6 +18,7 @@ import { TelemetryUtils } from "./utils/telemetry";
 import { Service } from "typedi";
 import { ResourcePlugins } from "../../solution/fx-solution/ResourcePluginContainer";
 import { isArmSupportEnabled } from "../../..";
+import "./v2";
 @Service(ResourcePlugins.SimpleAuthPlugin)
 export class SimpleAuthPlugin implements Plugin {
   name = "fx-resource-simple-auth";
