+++ packages/fx-core/src/plugins/solution/fx-solution/v2/scaffolding.ts
@@ -1,4 +1,4 @@
-import { v2, Inputs, FxError, Result, ok, err, Void } from "@microsoft/teamsfx-api";
+import { v2, Inputs, FxError, Result, ok, err, Void, AzureSolutionSettings } from "@microsoft/teamsfx-api";
 import { getStrings, isMultiEnvEnabled } from "../../../../common/tools";
 import {
   AzureResourceFunction,
@@ -7,7 +7,7 @@ import {
   TabOptionItem,
 } from "../question";
 import { executeConcurrently, NamedThunk } from "./executor";
-import { getAzureSolutionSettings, getSelectedPlugins } from "./utils";
+import { fillInSolutionSettings, getAzureSolutionSettings, getSelectedPlugins } from "./utils";
 import path from "path";
 import fs from "fs-extra";
 import { getTemplatesFolder } from "../../../..";
@@ -17,7 +17,10 @@ export async function scaffoldSourceCode(
   ctx: v2.Context,
   inputs: Inputs
 ): Promise<Result<Void, FxError>> {
-  const plugins = getSelectedPlugins(getAzureSolutionSettings(ctx));
+  const solutionSettings: AzureSolutionSettings = getAzureSolutionSettings(ctx);
+  const fillinRes = fillInSolutionSettings(solutionSettings, inputs);
+  if(fillinRes.isErr()) return err(fillinRes.error);
+  const plugins = getSelectedPlugins(solutionSettings);
 
   const thunks: NamedThunk<Void>[] = plugins
     .filter((plugin) => !!plugin.scaffoldSourceCode)
@@ -31,7 +34,6 @@ export async function scaffoldSourceCode(
     });
 
   const result = await executeConcurrently(thunks, ctx.logProvider);
-  const solutionSettings = getAzureSolutionSettings(ctx);
   if (result.isOk()) {
     const capabilities = solutionSettings.capabilities;
     const azureResources = solutionSettings.azureResources;
