+++ packages/fx-core/src/plugins/solution/fx-solution/v2/utils.ts
@@ -16,7 +16,7 @@ import {
 } from "@microsoft/teamsfx-api";
 import { LocalSettingsTeamsAppKeys } from "../../../../common/localSettingsConstants";
 import { SolutionError } from "../constants";
-import { HostTypeOptionAzure } from "../question";
+import { AzureResourceApim, AzureResourceFunction, AzureResourceSQL, AzureSolutionQuestionNames, BotOptionItem, HostTypeOptionAzure, MessageExtensionItem, TabOptionItem } from "../question";
 import {
   getActivatedResourcePlugins,
   getActivatedV2ResourcePlugins,
@@ -128,4 +128,48 @@ export function loadTeamsAppTenantIdForLocal(
       return ok(Void);
     }
   );
+}
+
+
+
+export function fillInSolutionSettings(solutionSettings: AzureSolutionSettings, answers: Inputs): Result<Void, FxError> {
+  const capabilities = (answers[AzureSolutionQuestionNames.Capabilities] as string[]) || [];
+  if (!capabilities || capabilities.length === 0) {
+    return err(
+      returnSystemError(
+        new Error("capabilities is empty"),
+        "Solution",
+        SolutionError.InternelError
+      )
+    );
+  }
+  let hostType = answers[AzureSolutionQuestionNames.HostType] as string;
+  if (capabilities.includes(BotOptionItem.id) || capabilities.includes(MessageExtensionItem.id))
+    hostType = HostTypeOptionAzure.id;
+  if (!hostType) {
+    return err(
+      returnSystemError(
+        new Error("hostType is undefined"),
+        "Solution",
+        SolutionError.InternelError
+      )
+    );
+  }
+  solutionSettings.hostType = hostType;
+  let azureResources: string[] | undefined;
+  if (hostType === HostTypeOptionAzure.id && capabilities.includes(TabOptionItem.id)) {
+    azureResources = answers[AzureSolutionQuestionNames.AzureResources] as string[];
+    if (azureResources) {
+      if (
+        (azureResources.includes(AzureResourceSQL.id) ||
+          azureResources.includes(AzureResourceApim.id)) &&
+        !azureResources.includes(AzureResourceFunction.id)
+      ) {
+        azureResources.push(AzureResourceFunction.id);
+      }
+    } else azureResources = [];
+  }
+  solutionSettings.azureResources = azureResources || [];
+  solutionSettings.capabilities = capabilities || [];
+  return ok(Void);
 }
